Issue #462 - wrong prototype for g_handler_sequential_number
`gulong` is a 8 bytes integer in GLib x64.

`handler_id` is generated by incrementing global `g_handler_sequential_number`
in GLib. This means that until it exceeds 4g, the bug is not visible.
When I open and close a working window in the product I'm working on,
counter jumps by ~20k. I understand that the value of 4g can be reached
realistically.

Also, afterall, it's a bug, let's fix it.

+++ bundles/org.eclipse.swt/Eclipse SWT PI/gtk/library/os.c
@@ -12356,7 +12356,7 @@ fail:
 
 #ifndef NO_g_1signal_1handler_1disconnect
 JNIEXPORT void JNICALL OS_NATIVE(g_1signal_1handler_1disconnect)
-	(JNIEnv *env, jclass that, jlong arg0, jint arg1)
+	(JNIEnv *env, jclass that, jlong arg0, jlong arg1)
 {
 	OS_NATIVE_ENTER(env, that, g_1signal_1handler_1disconnect_FUNC);
 	g_signal_handler_disconnect((gpointer)arg0, (gulong)arg1);
